Force refetching source if our cached copy will expire
If we set the `expiresAfter` option in metarefresh's config, we
effectively alter the inbound metadata. Thus when we're conditionally
GETting, we need to be aware that we've introduced a validUntil
attribute that the remote server may not know about. That means we
might need to force a refresh of the source even if the remote server
does not think it has been modified.